+++ include/AstuMath.h
@@ -21,6 +21,7 @@
 #include "Math/Vector2.h"
 #include "Math/Vector3.h"
 #include "Math/Matrix3.h"
+#include "Math/Matrix4.h"
 #include "Math/Transform2.h"
 #include "Math/Polygon.h"
 #include "Math/Line2.h"